# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Maryland Representatives Civic App API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /state/members/{memberId}:
    get:
      operationId: getStateMember
      tags: [state]
      summary: Get a state legislator's details
      parameters:
        - name: memberId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: State member details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StateMemberDetail"
  /state/members/{memberId}/bills:
    get:
      operationId: getStateMemberBills
      tags: [state]
      summary: Get bills sponsored or cosponsored by a state member
      parameters:
        - name: memberId
          in: path
          required: true
          schema:
            type: string
        - name: type
          in: query
          schema:
            type: string
            enum: [sponsored, cosponsored]
            default: sponsored
      responses:
        "200":
          description: State bills list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StateBillsListResponse"
  /state/members/{memberId}/votes:
    get:
      operationId: getStateMemberVotes
      tags: [state]
      summary: Get voting record for a state member
      parameters:
        - name: memberId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: State votes list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StateVotesListResponse"
  /state/bills:
    get:
      operationId: getStateBills
      tags: [state]
      summary: Get bills being considered in the Maryland legislature
      parameters:
        - name: chamber
          in: query
          schema:
            type: string
            enum: [upper, lower]
        - name: offset
          in: query
          schema:
            type: integer
            default: 0
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
      responses:
        "200":
          description: State bills list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StateBillsListResponse"
  /state/bills/{billId}:
    get:
      operationId: getStateBillDetail
      tags: [state]
      summary: Get detailed info about a specific state bill
      parameters:
        - name: billId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: State bill detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StateBillDetail"
...
components:
  schemas:
...
    SponsorInfo:
      type: object
      properties:
        name:
          type: string
        party:
          type: string
        state:
          type: string
        bioguideId:
          type: string
      required: [name]
    CommitteeRef:
      type: object
      properties:
        name:
          type: string
        chamber:
          type: string
        committeeCode:
          type: string
      required: [name]
    BillAction:
      type: object
      properties:
        date:
          type: string
        text:
          type: string
        type:
          type: string
      required: [date, text]
...
    StateMemberDetail:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        party:
          type: string
        chamber:
          type: string
        district:
          type: string
        email:
          type: string
        phone:
          type: string
        photoUrl:
          type: string
        openstatesUrl:
          type: string
      required: [id, name]
    StateBill:
      type: object
      properties:
        id:
          type: string
        identifier:
          type: string
        title:
          type: string
        session:
          type: string
        chamber:
          type: string
        status:
          type: string
        introducedDate:
          type: string
        latestAction:
          type: string
        sponsors:
          type: array
          items:
            type: string
        url:
          type: string
      required: [id, title]
    StateBillDetail:
      type: object
      properties:
        id:
          type: string
        identifier:
          type: string
        title:
          type: string
        session:
          type: string
        chamber:
          type: string
        status:
          type: string
        introducedDate:
          type: string
        summary:
          type: string
        sponsors:
          type: array
          items:
            $ref: "#/components/schemas/SponsorInfo"
        cosponsors:
          type: array
          items:
            $ref: "#/components/schemas/SponsorInfo"
        committees:
          type: array
          items:
            $ref: "#/components/schemas/CommitteeRef"
        actions:
          type: array
          items:
            $ref: "#/components/schemas/BillAction"
        votes:
          type: array
          items:
            $ref: "#/components/schemas/StateBillVote"
        url:
          type: string
      required: [id, title]
    StateBillVote:
      type: object
      properties:
        date:
          type: string
        chamber:
          type: string
        result:
          type: string
        yesCount:
          type: integer
        noCount:
          type: integer
        absentCount:
          type: integer
      required: [date, result]
    StateBillsListResponse:
      type: object
      properties:
        bills:
          type: array
          items:
            $ref: "#/components/schemas/StateBill"
        totalCount:
          type: integer
        offset:
          type: integer
      required: [bills]
    StateVoteRecord:
      type: object
      properties:
        billId:
          type: string
        billTitle:
          type: string
        billIdentifier:
          type: string
        date:
          type: string
        position:
          type: string
        result:
          type: string
      required: [billTitle, date, position]
    StateVotesListResponse:
      type: object
      properties:
        votes:
          type: array
          items:
            $ref: "#/components/schemas/StateVoteRecord"
      required: [votes]
...